Move extensions to custom_extensions.py
