+++ docs/conf.py
@@ -18,22 +18,17 @@
 # full list see the documentation:
 # http://www.sphinx-doc.org/en/master/config
 
+import datetime
 import os
 import sys
+
 sys.path.insert(0, os.path.abspath('.'))
 
-import qiskit_sphinx_theme
+from docs import custom_extensions
 from custom_directives import (IncludeDirective, GalleryItemDirective,
                                CustomGalleryItemDirective, CustomCalloutItemDirective,
                                CustomCardItemDirective)
 
-from distutils import dir_util
-import re
-import datetime
-import shutil
-import subprocess
-import tempfile
-import warnings
 
 # -- General configuration ---------------------------------------------------
 
@@ -263,111 +258,6 @@ autosummary_generate = True
 autosummary_generate_overwrite = False
 autoclass_content = 'both'
 
-# --- Custom Extensions -----------------------------------------------------
-
-# Elements with api doc sources
-qiskit_elements = ['qiskit-terra', 'qiskit-aer', 'qiskit-ibmq-provider']
-apidocs_exists = False
-apidocs_master = None
-
-
-def _get_current_versions(app):
-    versions = {}
-    setup_py_path = os.path.join(os.path.dirname(app.srcdir), 'setup.py')
-    with open(setup_py_path, 'r') as fd:
-        setup_py = fd.read()
-        for package in qiskit_elements:
-            version_regex = re.compile(package + '[=|>]=(.*)\"')
-            match = version_regex.search(setup_py)
-            if match:
-                ver = match[1]
-                versions[package] = ver
-    return versions
-
-
-def _install_from_master():
-    for package in qiskit_elements + ['qiskit-ignis']:
-        github_url = 'git+https://github.com/Qiskit/%s' % package
-        cmd = [sys.executable, '-m', 'pip', 'install', '-U', github_url]
-        subprocess.run(cmd)
-
-
-def _git_copy(package, sha1, api_docs_dir):
-    try:
-        with tempfile.TemporaryDirectory() as temp_dir:
-            github_source = 'https://github.com/Qiskit/%s' % package
-            subprocess.run(['git', 'clone', github_source, temp_dir],
-                           capture_output=True)
-            subprocess.run(['git', 'checkout', sha1], cwd=temp_dir,
-                           capture_output=True)
-            dir_util.copy_tree(
-                os.path.join(temp_dir, 'docs', 'apidocs'),
-                api_docs_dir)
-
-    except FileNotFoundError:
-        warnings.warn('Copy from git failed for %s at %s, skipping...' %
-                      (package, sha1), RuntimeWarning)
-
-
-def load_api_sources(app):
-    api_docs_dir = os.path.join(app.srcdir, 'apidoc')
-    if os.getenv('DOCS_FROM_MASTER'):
-        global apidocs_master
-        apidocs_master = tempfile.mkdtemp()
-        shutil.move(api_docs_dir, apidocs_master)
-        _install_from_master()
-        for package in qiskit_elements:
-            _git_copy(package, 'HEAD', api_docs_dir)
-        return
-    elif os.path.isdir(api_docs_dir):
-        global apidocs_exists
-        apidocs_exists = True
-        warnings.warn('docs/apidocs already exists skipping source clone')
-        return
-    meta_versions = _get_current_versions(app)
-    for package in qiskit_elements:
-        _git_copy(package, meta_versions[package], api_docs_dir)
-
-def load_tutorials(app):
-    tutorials_dir = os.path.join(app.srcdir, 'tutorials')
-    try:
-        with tempfile.TemporaryDirectory() as temp_dir:
-            github_source = 'https://github.com/Qiskit/qiskit-tutorials'
-            subprocess.run(['git', 'clone', github_source, temp_dir],
-                           capture_output=True)
-            dir_util.copy_tree(
-                os.path.join(temp_dir, 'tutorials'),
-                tutorials_dir)
-    except FileNotFoundError:
-        warnings.warn('Copy from git failed for %s at %s, skipping...' %
-                      (package, sha1), RuntimeWarning)
-
-
-def clean_api_source(app, exc):
-    api_docs_dir = os.path.join(app.srcdir, 'apidoc')
-    global apidocs_exists
-    global apidocs_master
-    if apidocs_exists:
-        return
-    elif apidocs_master:
-        shutil.rmtree(api_docs_dir)
-        shutil.move(os.path.join(apidocs_master, 'apidoc'), api_docs_dir)
-        return
-    shutil.rmtree(api_docs_dir)
-
-
-def clean_tutorials(app, exc):
-    tutorials_dir = os.path.join(app.srcdir, 'tutorials')
-    shutil.rmtree(tutorials_dir)
-
-def deprecate_ibmq_provider(app, docname, source):
-    message = """.. warning::
-       The package ``qiskit-ibmq-provider`` is being deprecated and its repo is going to be
-       archived soon. Please transition to the new packages. More information in
-       https://ibm.biz/provider_migration_guide\n\n"""
-    if 'apidoc/ibmq' in docname or 'qiskit.providers.ibmq' in docname:
-        source[0] = message + source[0]
-
 # -- Extension configuration -------------------------------------------------
 
 def setup(app):
@@ -376,10 +266,10 @@ def setup(app):
     app.add_directive('customgalleryitem', CustomGalleryItemDirective)
     app.add_directive('customcarditem', CustomCardItemDirective)
     app.add_directive('customcalloutitem', CustomCalloutItemDirective)
-    load_api_sources(app)
-    load_tutorials(app)
+    custom_extensions.load_api_sources(app)
+    custom_extensions.load_tutorials(app)
     app.setup_extension('versionutils')
     app.add_css_file('css/theme-override.css')
-    app.connect('build-finished', clean_api_source)
-    app.connect('build-finished', clean_tutorials)
-    app.connect('source-read', deprecate_ibmq_provider)
+    app.connect('build-finished', custom_extensions.clean_api_source)
+    app.connect('build-finished', custom_extensions.clean_tutorials)
+    app.connect('source-read', custom_extensions.deprecate_ibmq_provider)
